DRY up project page serverside props
